DELETING FILE FROM PATH WHEN PRODUCT IS DELETED WORKS FINE

+++ controllers/admin.js
@@ -1,6 +1,10 @@
 const mongodb = require('mongodb');
 const Product = require('../models/product');
 
+
+const fileHelper = require('../util/file');
+
+
 //importing express validator
 const { validationResult } = require('express-validator');
 
@@ -158,6 +162,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
 
       if(updatedImage) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = updatedImage.path;
       }
 
@@ -203,17 +208,21 @@ exports.getProductsList = (req, res, next) => {
 //This deletes product by id 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products-list');
     })
     .catch(err => {
-      //console.log(err)
-      //res.redirect('/500');
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error);}
-      );
-
+      return next(error);
+    });
 };
